+++ core-api/src/main/java/io/ssafy/p/k13c103/coreapi/domain/member/controller/MemberController.java
@@ -3,7 +3,7 @@ package io.ssafy.p.k13c103.coreapi.domain.member.controller;
 import io.ssafy.p.k13c103.coreapi.common.jsend.JSend;
 import io.ssafy.p.k13c103.coreapi.domain.member.dto.MemberRequestDto;
 import io.ssafy.p.k13c103.coreapi.domain.member.dto.MemberResponseDto;
-import io.ssafy.p.k13c103.coreapi.domain.member.MemberService;
+import io.ssafy.p.k13c103.coreapi.domain.member.service.MemberService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
