+++ core-api/src/main/java/io/ssafy/p/k13c103/coreapi/domain/member/service/MemberService.java
@@ -1,4 +1,4 @@
-package io.ssafy.p.k13c103.coreapi.domain.member;
+package io.ssafy.p.k13c103.coreapi.domain.member.service;
 
 import io.ssafy.p.k13c103.coreapi.domain.member.dto.MemberRequestDto;
 import io.ssafy.p.k13c103.coreapi.domain.member.dto.MemberResponseDto;
